Atualizações: - Variáveis globais convertidas para uint_t para uso de ponto fixo - Função de recebimento de pedidos implementada com armazenamento via ponteiros - Função de transmissão criada, enviando dados globais e convertendo para ASCII

+++ Trabalho_final.X/globals.c
@@ -12,8 +12,8 @@ volatile uint8_t andar_atual = 0;       // Inicia no Térreo (0)
 volatile uint8_t andar_destino = 0;     // Sem destino inicial
 volatile uint8_t estado_motor = MOTOR_PARADO; 
 volatile uint8_t posicao_mm = 0;        // 0 mm (Térreo)
-volatile float velocidade_atual = 0.0;  // 0.0 mm/s
-volatile float temperatura_ponte = 0.0; // 0.0 °C
+volatile uint8_t velocidade_atual = 0;  // 0.0*10 mm/s
+volatile uint8_t temperatura_ponte = 0; // 0.0*10 °C
 
 // Vetor de Solicitações (Botões apertados)
 // [0]=Térreo, [1]=1ºAndar, [2]=2ºAndar, [3]=3ºAndar
